% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleTantan.R
\docType{data}
\name{exampleTantan}
\alias{exampleTantan}
\title{Example tantan data}
\format{
A \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}} object.
}
\usage{
exampleTantan
}
\description{
Example object loaded from the output of the \code{tantan -f4} program after
reordering and filterning columns with
\verb{awk 'OFS="\\t" \{print $1, $2, $3, $6, $5, "."\}}
}
\author{
Charles Plessy
}
\keyword{datasets}
