% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tantan.R
\name{load_tantan}
\alias{load_tantan}
\title{Load tantan results}
\usage{
load_tantan(file)
}
\arguments{
\item{file}{The output file to load}
}
\value{
A \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}} object.
}
\description{
Load the output of \verb{tantan -f4 awk 'OFS="\\t" \{print $1, $2, $3, $6, $5, "."\}}.
}
