% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTantan.R
\name{loadTantan}
\alias{loadTantan}
\title{Load tantan results}
\usage{
loadTantan(file, genome.file = NA, genome.name = NA)
}
\arguments{
\item{file}{The output file to load}

\item{genome.file}{Path to the genome file that was given to \code{tantan}}

\item{genome.name}{A genome name for the \code{seqinfo} of the returned \code{GRanges}}
}
\value{
A \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}} object.
}
\description{
Load the output of \verb{tantan -f4 | awk 'OFS="\\t" \{print $1, $2, $3, $6, $5, "."\}}.
}
\details{
If given the genome file that \code{tantan} used, it will use it to create a
\code{seqinfo} object for the returned \code{GRanges} object
}
\examples{
loadTantan(system.file('extdata/example.bed.xz', package="changchang"))
}
