% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_tandem_seq.R
\name{tandemNormalForm}
\alias{tandemNormalForm}
\alias{tandemNormalForm,character-method}
\alias{tandemNormalForm,factor-method}
\alias{tandemNormalForm,DNAStringSet-method}
\alias{tandemNormalForm,DNAString-method}
\title{Normal form of tandem sequence}
\usage{
tandemNormalForm(seq, strand = c("forward", "reverse", "both"))

\S4method{tandemNormalForm}{character}(seq, strand = c("forward", "reverse", "both"))

\S4method{tandemNormalForm}{factor}(seq, strand = c("forward", "reverse", "both"))

\S4method{tandemNormalForm}{DNAStringSet}(seq, strand = c("forward", "reverse", "both"))

\S4method{tandemNormalForm}{DNAString}(seq, strand = c("forward", "reverse", "both"))
}
\arguments{
\item{seq}{tandem repeat base unit sequence}

\item{strand}{With strand to consider}
}
\value{
Returns normal forms of the tandem unit sequences, in the same class
as the input.
}
\description{
Computes the lexicographically smallest rotation of a tandem repeat base
unit.
}
\examples{
telomeres <- c("TAAGGG", "CCCTTA")
tandemNormalForm(telomeres)
tandemNormalForm(telomeres, strand="reverse")
tandemNormalForm(telomeres, strand="both")
tandemNormalForm(factor(telomeres), strand="both") # Should return a factor
tandemNormalForm(Biostrings::DNAStringSet(telomeres), strand="both") # Should return a DNAStringSet

}
\author{
Charles Plessy
}
