% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telomeres.R
\name{telomeres}
\alias{telomeres}
\title{Search for telomeres}
\usage{
telomeres(gr, motif = "TTAGGG", narrow = c("no", "ends"))
}
\arguments{
\item{gr}{A \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}} object produced by the \code{\link{loadTandems}}
function}

\item{motif}{The telomeric sequence motif, at the end of the plus strand,
like in TeloBase and the Telomerase Database.}

\item{narrow}{Narrow the search on the ends.}
}
\description{
Searches for telomeric repeats in the whole list of tandem repeats, or
narrowing the search on the ends of sequence features.
}
\examples{
telomeres(exampleTantan)
telomeres(exampleTantan, narrow="ends")
}
\author{
Charles Plessy
}
